Moved files, fixed headers, created build file
