+++ CondFormats/SiPixelCommon/BuildFile.xml
@@ -1,0 +1,7 @@
+<use   name="FWCore/Framework"/>
+<use   name="FWCore/Utilities"/>
+<use   name="boost"/>
+
+<export>
+  <lib   name="1"/>
+</export>
