+++ CondFormats/SiPixelCommon/interface/SiPixelGenError.h
