+++ RecoLocalTracker/SiPixelRecHits/interface/PixelCPETemplateReco.h
@@ -14,7 +14,7 @@
 
 
 #ifndef SI_PIXEL_TEMPLATE_STANDALONE
-#include "RecoLocalTracker/SiPixelRecHits/interface/SiPixelTemplate.h"
+#include "CondFormats/SiPixelCommon/interface/SiPixelTemplate.h"
 #else
 #include "SiPixelTemplate.h"
 #endif
