+++ RecoLocalTracker/SiPixelRecHits/interface/SiPixelTemplateReco.h
@@ -54,8 +54,8 @@
 #define SiPixelTemplateReco_h 1
 
 #ifndef SI_PIXEL_TEMPLATE_STANDALONE
-#include "RecoLocalTracker/SiPixelRecHits/interface/SiPixelTemplateDefs.h"
-#include "RecoLocalTracker/SiPixelRecHits/interface/SiPixelTemplate.h"
+#include "CondFormats/SiPixelCommon/interface/SiPixelTemplateDefs.h"
+#include "CondFormats/SiPixelCommon/interface/SiPixelTemplate.h"
 #else
 #include "SiPixelTemplateDefs.h"
 #include "SiPixelTemplate.h"
